+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/components/projectoverview/ProjectOverviewView.groovy
@@ -1,14 +1,20 @@
 package life.qbic.portal.sampletracking.components.projectoverview
 
 import com.vaadin.data.provider.ListDataProvider
+import com.vaadin.event.selection.SingleSelectionEvent
+import com.vaadin.icons.VaadinIcons
+import com.vaadin.server.FileDownloader
+import com.vaadin.server.StreamResource
+import com.vaadin.shared.ui.ContentMode
 import com.vaadin.shared.ui.grid.HeightMode
-import com.vaadin.ui.Grid
-import com.vaadin.ui.Label
-import com.vaadin.ui.VerticalLayout
+import com.vaadin.ui.*
 import com.vaadin.ui.themes.ValoTheme
+import life.qbic.portal.sampletracking.Constants
+import life.qbic.portal.sampletracking.communication.notification.NotificationService
+import life.qbic.portal.sampletracking.components.projectoverview.download.DownloadProjectController
 
 /**
- * <h1>This class generates the layout for the ProductOverview use case</h1>
+ * <b>This class generates the layout for the ProductOverview use case</b>
  *
  * <p>This view will be the entry point for the user and provides an overview of her projects. And the projects overall status.
  * From here, the user can navigate to the StatusOverview.</p>
@@ -18,29 +24,106 @@ import com.vaadin.ui.themes.ValoTheme
 */
 class ProjectOverviewView extends VerticalLayout{
 
+    private final ProjectOverviewViewModel viewModel
+    private final DownloadProjectController downloadProjectController
+    private final NotificationService notificationService
+
     private Label titleLabel
-    private ProjectOverviewViewModel viewModel
     private Grid<ProjectSummary> projectGrid
 
     final static int MAX_CODE_COLUMN_WIDTH = 400
     final static int MAX_STATUS_COLUMN_WIDTH = 200
+    private FileDownloader fileDownloader
 
-    ProjectOverviewView(ProjectOverviewViewModel viewModel){
+    ProjectOverviewView(NotificationService notificationService, ProjectOverviewViewModel viewModel, DownloadProjectController downloadProjectController){
+        this.notificationService = notificationService
         this.viewModel = viewModel
+        this.downloadProjectController = downloadProjectController
 
         initLayout()
-        fillGrid()
     }
 
     private void initLayout(){
         titleLabel = new Label("Project Overview")
         titleLabel.addStyleName(ValoTheme.LABEL_LARGE)
-        projectGrid = new Grid<>()
-
-        this.addComponents(titleLabel, projectGrid)
+        setupProjects()
+        Component component = setupProjectSpecificButtons()
+        this.addComponents(titleLabel,component, projectGrid)
     }
 
-    private void fillGrid(){
+    private Button setUpLinkButton(){
+        Button button = new Button()
+        button.setIcon(VaadinIcons.QUESTION_CIRCLE)
+        button.setStyleName(ValoTheme.BUTTON_ICON_ONLY + " " + ValoTheme.BUTTON_SMALL + " square")
+        button.setDescription("A manifest is a text file used by a client application (e.g. <a href=\"https://github.com/qbicsoftware/postman-cli\" target=\"_blank\">qpostman</a>) to download selected files of interest. <br>" +
+                "Use <a href=\"https://github.com/qbicsoftware/postman-cli\" target=\"_blank\">qpostman</a> to download the data.", ContentMode.HTML)
+
+        button.addClickListener({
+            getUI().getPage().open(
+                    "https://github.com/qbicsoftware/postman-cli#provide-a-file-with-several-qbic-ids",
+                    "_blank")
+        })
+        return button
+    }
+
+    private Button setupDownloadButton() {
+        Button downloadManifestAction = new Button("Download Manifest", VaadinIcons.DOWNLOAD)
+        viewModel.addPropertyChangeListener("generatedManifest", {
+            if (isDownloadAvailable()) {
+                this.fileDownloader = new FileDownloader(new StreamResource({viewModel.getManifestInputStream()}, "manifest.txt"))
+                this.fileDownloader.extend(downloadManifestAction)
+            } else {
+                if (this.fileDownloader) {
+                    if (downloadManifestAction.extensions.contains(fileDownloader)) {
+                        downloadManifestAction.removeExtension(this.fileDownloader)
+                    }
+                }
+            }
+        })
+        enableWhenDownloadIsAvailable(downloadManifestAction)
+        return downloadManifestAction
+    }
+
+    private void setupProjects(){
+        projectGrid = new Grid<>()
+        fillProjectsGrid()
+        projectGrid.setSelectionMode(Grid.SelectionMode.SINGLE)
+        projectGrid.addSelectionListener({
+            if (it instanceof SingleSelectionEvent<ProjectSummary>) {
+                clearProjectSelection()
+                it.getSelectedItem().ifPresent(this::selectProject)
+            }
+        })
+        viewModel.addPropertyChangeListener("selectedProject", {
+            Optional<ProjectSummary> modelSelection = Optional.ofNullable(viewModel.selectedProject)
+            Optional<ProjectSummary> viewSelection = projectGrid.getSelectionModel().getFirstSelectedItem()
+            modelSelection.ifPresent({
+                if (viewSelection.isPresent()) {
+                    if (viewSelection.get() == modelSelection.get()) {
+                        // do nothing
+                    } else {
+                        projectGrid.getSelectionModel().deselectAll()
+                        projectGrid.select(modelSelection.get())
+                    }
+                } else {
+                    projectGrid.select(modelSelection.get())
+                }
+            })
+
+        })
+    }
+
+    private void selectProject(ProjectSummary projectSummary) {
+        viewModel.selectedProject = projectSummary
+        tryToDownloadManifest()
+    }
+
+    private void clearProjectSelection() {
+        viewModel.selectedProject = null
+        viewModel.generatedManifest = null
+    }
+
+    private void fillProjectsGrid() {
         projectGrid.addColumn({ it.code})
                 .setCaption("Project Code").setId("ProjectCode").setMaximumWidth(MAX_CODE_COLUMN_WIDTH)
         projectGrid.addColumn({ it.title })
@@ -49,6 +132,8 @@ class ProjectOverviewView extends VerticalLayout{
                 .setCaption("Samples Received").setId("SamplesReceived")
         projectGrid.addColumn({it.samplesQcFailed})
                 .setCaption("Samples Failed QC").setId("SamplesFailedQc")
+        projectGrid.addColumn({it.sampleDataAvailable})
+                .setCaption("Data Available").setId("SampleDataAvailable")
         setupDataProvider()
         //specify size of grid and layout
         projectGrid.setWidthFull()
@@ -58,6 +143,8 @@ class ProjectOverviewView extends VerticalLayout{
                 .setMaximumWidth(MAX_STATUS_COLUMN_WIDTH).setExpandRatio(1)
         projectGrid.getColumn("SamplesFailedQc")
                 .setMaximumWidth(MAX_STATUS_COLUMN_WIDTH).setExpandRatio(1)
+        projectGrid.getColumn("SampleDataAvailable")
+                .setMaximumWidth(MAX_STATUS_COLUMN_WIDTH).setExpandRatio(1)
         projectGrid.setHeightMode(HeightMode.ROW)
     }
 
@@ -65,4 +152,40 @@ class ProjectOverviewView extends VerticalLayout{
         def dataProvider = new ListDataProvider(viewModel.projectOverviews)
         projectGrid.setDataProvider(dataProvider)
     }
+
+    private AbstractComponent setupProjectSpecificButtons() {
+        HorizontalLayout buttonBar = new HorizontalLayout()
+        buttonBar.setMargin(false)
+        Button postmanLink = setUpLinkButton()
+        Button downloadManifestAction = setupDownloadButton()
+        buttonBar.addComponents(downloadManifestAction, postmanLink)
+        buttonBar.setComponentAlignment(postmanLink, Alignment.MIDDLE_CENTER)
+        return buttonBar
+    }
+
+    private void tryToDownloadManifest() {
+        try {
+            Optional.ofNullable(viewModel.selectedProject).filter({it.sampleDataAvailable > 0 }).ifPresent({
+                String projectCode = it.getCode()
+                downloadProjectController.downloadProject(projectCode)
+            })
+        } catch (IllegalArgumentException illegalArgument) {
+            notificationService.publishFailure("Manifest Download failed due to: ${illegalArgument.getMessage()}")
+        } catch (Exception ignored) {
+            notificationService.publishFailure("Manifest Download failed for unknown reasons. ${Constants.CONTACT_HELPDESK}")
+        }
+    }
+
+    private void enableWhenDownloadIsAvailable(Component component) {
+        component.enabled = isDownloadAvailable()
+        viewModel.addPropertyChangeListener("generatedManifest") {
+            component.enabled = isDownloadAvailable()
+        }
+    }
+
+    private boolean isDownloadAvailable() {
+        boolean isAvailable = viewModel.generatedManifest as boolean
+        return isAvailable
+    }
+
 }