+++ sample-tracking-status-overview-app/src/main/groovy/life/qbic/portal/sampletracking/datasources/samples/SamplesDbConnector.groovy
@@ -23,7 +23,6 @@ import java.sql.ResultSet
  */
 @Log4j2
 class SamplesDbConnector implements CountSamplesDataSource, DownloadSamplesDataSource {
-
     private final ConnectionProvider connectionProvider
 
     /**