+++ sample-tracking-status-overview-domain/src/main/groovy/life/qbic/business/samples/count/CountSamples.groovy
@@ -45,7 +45,7 @@ class CountSamples implements CountSamplesInput{
       output.countedReceivedSamples(projectCode, sampleStatuses.size(), receivedAmount)
     } catch (DataSourceException dataSourceException) {
       output.failedExecution(dataSourceException.getMessage())
-    } catch (Exception e) {
+    } catch (Exception ignored) {
       output.failedExecution("Could not count received samples.")
     }
   }
@@ -62,8 +62,19 @@ class CountSamples implements CountSamplesInput{
       sampleStatuses = dataSource.fetchSampleStatusesForProject(projectCode)
       int receivedAmount = sampleStatuses.findAll { it == Status.SAMPLE_QC_FAIL }.size()
       output.countedFailedQcSamples(projectCode, sampleStatuses.size(), receivedAmount)
-    }catch (Exception e) {
+    }catch (Exception ignored) {
       output.failedExecution("Could not count failed qc samples.")
+    }
+  }
+
+  @Override
+  void countAvailableDataSamples(String projectCode) {
+    try {
+      sampleStatuses = dataSource.fetchSampleStatusesForProject(projectCode)
+      int availableData = countSamplesFromStatus(Status.DATA_AVAILABLE)
+      output.countedAvailableSampleData(projectCode, sampleStatuses.size(), availableData)
+    } catch (Exception e) {
+      output.failedExecution(e.getMessage())
     }
   }